# specs/002-builder-modelli/contracts/builder-modelli-api.openapi.yaml
openapi: 3.0.3
info:
  title: GEMODO - Builder Modelli
  version: 0.6.0
  description: >
    Creazione e ciclo di vita di modello/versione (007 tasks.md T006). Sostituisce
    integralmente la 0.2.0: quella descriveva tipi documento/categorie come CRUD
    locale del builder, un design pre-010 mai implementato cosi'. Oggi tipi
    documento/categorie sono letti in sola lettura dalla discovery live
    dell'integrazione connessa (vedi integrazioni-api.openapi.yaml, gia'
    pubblicato) - questo contratto copre solo le route realmente implementate in
    backend/app/builder/api.py che quel contratto non gia' copre: lettura
    struttura disponibile per un tipo documento gia' associato, creazione
    modello/versione, transizioni di stato BOZZA -> IN_REVISIONE -> APPROVATO ->
    PUBBLICATO. Nessuna route di archiviazione/sospensione/bozza-derivata esiste
    ancora (Nota Implementazione 2026-09-16 in tasks.md).
servers:
  - url: /api/v1/builder
security:
  - KeycloakBearer: []
paths:
...
  /modelli:
    get:
      operationId: listaModelliBuilder
      summary: Modelli e versioni del contesto autorizzato, comprese bozze
      description: Ordinati per creazione decrescente e id. Non richiede discovery online. Nessuna URL amministrativa esposta.
      parameters:
        - name: codice_contesto
          in: query
          required: true
          schema: {type: string, minLength: 1, maxLength: 64}
        - name: offset
          in: query
          schema: {type: integer, minimum: 0, default: 0}
        - name: limit
          in: query
          schema: {type: integer, minimum: 1, maximum: 100, default: 50}
        # Filtri lato server (007 FR-028). Stanno qui e non nel client perche'
        # la lista e' paginata: filtrare dopo `limit` restituirebbe la pagina in
        # mano invece dell'insieme. Un parametro assente significa "non filtrare".
        - name: codice_tipo_documento
          in: query
          schema: {type: string, minLength: 1, maxLength: 128}
        - name: integrazione_id
          in: query
          schema: {type: string, format: uuid}
        - name: codice_tipologia
          in: query
          schema: {type: string, minLength: 1, maxLength: 128}
        - name: codice_categoria
          in: query
          schema: {type: string, minLength: 1, maxLength: 128}
        - name: lingua
          in: query
          schema: {type: string, enum: [IT, EN]}
        - name: livello_professionale
          in: query
          description: Il token TUTTI seleziona i modelli senza livello, dove la colonna e' nulla.
          schema: {type: string, minLength: 1, maxLength: 64}
        - name: variante
          in: query
          schema: {type: string, minLength: 1, maxLength: 64}
        - name: stato_versione
          in: query
          description: Seleziona i modelli che hanno almeno una versione in questo stato.
          schema:
            type: string
            enum: [BOZZA, IN_REVISIONE, APPROVATO, PUBBLICATO, ARCHIVIATO, SOSPESO]
        - name: ricerca
          in: query
          description: >
            Sottostringa su nome e codice, applicata prima della paginazione.
            Usa ILIKE e non un indice: adeguata a modelli nell'ordine delle
            centinaia, da rivedere se crescessero di ordini di grandezza.
          schema: {type: string, minLength: 1, maxLength: 200}
      responses:
        '200':
          description: Pagina modelli filtrata, vuota se nessun modello corrisponde
          content:
            application/json:
              schema: {type: array, items: {$ref: '#/components/schemas/ModelloGestione'}}
              example: []
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '400': {$ref: '#/components/responses/InvalidInput'}
    post:
      operationId: creaModello
      summary: Crea un modello in BOZZA a partire da un percorso o categoria/tipologia risolti in discovery
      description: >
        Autorizzazione per il singolo contesto del tipo documento target
        (verify_scrittura_su_contesto), mai sull'unione dei contesti del token.
      x-required-role: GEMODO_MODELLI_GESTORE (nel contesto del tipo documento target)
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/CreaModelloRequest'}
            example:
              codice_tipo_documento: BANDO_CONCORSO
              integrazione_id: 00000000-0000-0000-0000-000000000001
              percorso_categorizzazione: [TD, RICERCATORE]
              livello_professionale: VI
              lingua: IT
      responses:
        '201':
          description: Modello creato in stato BOZZA (senza ancora una versione)
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Modello'}
        '400': {$ref: '#/components/responses/InvalidInput'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
...
  /modelli/{modelloId}:
    parameters:
      - {$ref: '#/components/parameters/ModelloId'}
    get:
      operationId: getModelloDettaglio
      summary: Dettaglio modello con il contratto dati, leggibile anche su una BOZZA
      description: >
        Alimenta l'anteprima del modello nel builder (schermata 2b in versione
        ridotta, 007). A differenza di /catalogo/modelli/{modelloVersioneId}/campi-richiesti,
        che risponde solo su versioni pubblicate, qui i campi sono leggibili anche
        su una versione in BOZZA. Stesso isolamento per contesto della lista.
      x-required-role: GEMODO_MODELLI_GESTORE
      responses:
        '200':
          description: Modello con versioni e campi del contratto
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ModelloDettaglio'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
    delete:
      operationId: eliminaModello
      summary: Elimina logicamente un modello conservando versioni, audit e PDF
      parameters:
        - name: modelloId
          in: path
          required: true
          schema: {type: string, format: uuid}
      responses:
        '204': {description: 'Modello eliminato, versioni pubblicate archiviate'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
...
components:
  securitySchemes:
    KeycloakBearer: {type: http, scheme: bearer, bearerFormat: JWT}
  parameters:
...
    ModelloId:
      name: modelloId
      in: path
      required: true
      schema: {type: string, format: uuid}
...
  schemas:
...
    ModelloGestione:
      type: object
      additionalProperties: false
      required: [id, public_id, codice, nome, codice_tipo_documento, codice_categoria, codice_tipologia, percorso_categorizzazione, variante, dimensioni, lingua, livello_professionale, codice_contesto, integrazione_id, created_at, versioni]
      properties:
        id: {type: string, format: uuid}
        public_id: {type: integer, nullable: true}
        codice: {type: string}
        nome: {type: string}
        codice_tipo_documento: {type: string}
        codice_categoria: {type: string}
        codice_tipologia: {type: string, nullable: true}
        percorso_categorizzazione: {type: array, items: {type: string}}
        variante: {type: string}
        dimensioni:
          type: object
          additionalProperties: {type: string, minLength: 1}
        lingua: {type: string, enum: [IT, EN], nullable: true}
        livello_professionale: {type: string, nullable: true}
        derivato_da_modello_id: {type: string, format: uuid, nullable: true}
        codice_contesto: {type: string}
        integrazione_id: {type: string, format: uuid, nullable: true}
        created_at: {type: string, format: date-time}
        versioni: {type: array, items: {$ref: '#/components/schemas/Versione'}}
...
    CampoContratto:
      description: >
        Il campo come lo legge l'anteprima del builder: metadati completi, a
        differenza di CampoVersione che e' la forma di richiesta alla creazione.
      type: object
      additionalProperties: false
      required: [codice, etichetta, tipo, lingua, obbligatorio, ordine]
      properties:
        codice: {type: string}
        etichetta: {type: string}
        tipo: {type: string}
        lingua: {type: string, enum: [IT, EN]}
        obbligatorio: {type: boolean}
        ordine: {type: integer}
        descrizione: {type: string, nullable: true}
    VersioneDettaglio:
      allOf:
        - $ref: '#/components/schemas/Versione'
        - type: object
          required: [campi]
          properties:
            campi: {type: array, items: {$ref: '#/components/schemas/CampoContratto'}}
    ModelloDettaglio:
      type: object
      additionalProperties: false
      required: [id, public_id, codice, nome, codice_tipo_documento, codice_categoria, codice_tipologia, percorso_categorizzazione, variante, dimensioni, lingua, livello_professionale, codice_contesto, integrazione_id, created_at, dimensioni_non_disponibili, versioni]
      properties:
        id: {type: string, format: uuid}
        public_id: {type: integer, nullable: true}
        codice: {type: string}
        nome: {type: string}
        codice_tipo_documento: {type: string}
        codice_categoria: {type: string}
        codice_tipologia: {type: string, nullable: true}
        percorso_categorizzazione: {type: array, items: {type: string}}
        variante: {type: string}
        dimensioni:
          type: object
          additionalProperties: {type: string, minLength: 1}
        lingua: {type: string, enum: [IT, EN], nullable: true}
        livello_professionale: {type: string, nullable: true}
        derivato_da_modello_id: {type: string, format: uuid, nullable: true}
        codice_contesto: {type: string}
        integrazione_id: {type: string, format: uuid, nullable: true}
        created_at: {type: string, format: date-time}
        dimensioni_non_disponibili:
          type: array
          description: Dimensioni salvate sul modello ma non piu' dichiarate dalla foglia live.
          items: {type: string}
        versioni: {type: array, items: {$ref: '#/components/schemas/VersioneDettaglio'}}
    CreaModelloRequest:
      type: object
      additionalProperties: false
      required: [codice_tipo_documento]
      description: >
        percorso_categorizzazione XOR codice_categoria (+ codice_tipologia
        opzionale se il percorso e' altrimenti ambiguo) - vedi verifica_selezione
        in backend/app/builder/schemas.py.
      properties:
        codice_tipo_documento: {type: string, minLength: 1}
        integrazione_id:
          type: string
          format: uuid
          description: Integrazione selezionata; risolve il tipo per sorgente e codice, senza fallback legacy. Omettere solo per chiamate legacy non ambigue.
        percorso_categorizzazione:
          type: array
          nullable: true
          minItems: 1
          maxItems: 64
          items: {type: string, minLength: 1}
        codice_categoria: {type: string, nullable: true}
        codice_tipologia: {type: string, nullable: true}
        dimensioni:
          type: object
          default: {}
          additionalProperties: {type: string, minLength: 1}
        lingua: {type: string, enum: [IT, EN]}
        livello_professionale:
          type: string
          nullable: true
          description: Valore dichiarato in livelli_possibili; null significa tutti i livelli della foglia.
...
    Modello:
      type: object
      additionalProperties: false
      required: [id, public_id, codice, nome, codice_tipo_documento, codice_categoria, codice_tipologia, percorso_categorizzazione, variante, lingua, livello_professionale]
      properties:
        id: {type: string, format: uuid}
        public_id: {type: integer, nullable: true}
        codice: {type: string}
        nome: {type: string}
        codice_tipo_documento: {type: string}
        codice_categoria: {type: string}
        codice_tipologia: {type: string, nullable: true}
        percorso_categorizzazione: {type: array, items: {type: string}}
        variante: {type: string}
        lingua: {type: string, enum: [IT, EN]}
        livello_professionale: {type: string, nullable: true}
        derivato_da_modello_id: {type: string, format: uuid, nullable: true}
...
    Versione:
      type: object
      additionalProperties: false
      required: [id, public_id, modello_id, numero_versione, stato, pubblicato_at]
      properties:
        id: {type: string, format: uuid}
        public_id: {type: integer, nullable: true}
        modello_id: {type: string, format: uuid}
        numero_versione: {type: integer}
        stato: {type: string, enum: [BOZZA, IN_REVISIONE, APPROVATO, PUBBLICATO, ARCHIVIATO, SOSPESO]}
        pubblicato_at: {type: string, format: date-time, nullable: true}
    Errore:
      type: object
      additionalProperties: false
      required: [codice, messaggio]
      properties:
        codice: {type: string}
        messaggio: {type: string}
  responses:
    Unauthorized:
      description: JWT assente o non valido
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Errore'}
          example: {codice: ACCESSO_NON_AUTENTICATO, messaggio: Autenticazione richiesta}
    Forbidden:
      description: Ruolo/contesto non autorizzato (mai sull'unione dei contesti del token)
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Errore'}
          example: {codice: ACCESSO_NON_AUTORIZZATO, messaggio: Operazione non autorizzata}
    NotFound:
      description: Modello/versione inesistente, tipo documento non trovato, o ramo non piu' disponibile in discovery
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Errore'}
          examples:
            modello: {value: {codice: MODELLO_NON_TROVATO, messaggio: Modello non trovato}}
            versione: {value: {codice: MODELLO_VERSIONE_NON_TROVATO, messaggio: Versione non trovata}}
            ramo: {value: {codice: CONTESTO_NON_VALIDO, messaggio: Il ramo del modello non e' piu' disponibile}}
    Conflict:
      description: Integrazione non connessa, tipo documento gia' di un'altra integrazione o transizione di stato non valida
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Errore'}
          examples:
            nonConnessa: {value: {codice: INTEGRAZIONE_NON_CONNESSA, messaggio: Integrazione non connessa}}
            transizione: {value: {codice: TRANSIZIONE_STATO_NON_VALIDA, messaggio: Transizione di stato non valida}}
            altraIntegrazione: {value: {codice: TIPO_DOCUMENTO_ALTRA_INTEGRAZIONE, messaggio: Tipo documento gia' associato a un'altra integrazione dello stesso contesto}}
    InvalidInput:
      description: Selezione percorso/categoria incoerente o campi non ammessi/duplicati
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Errore'}
          examples:
            percorso: {value: {codice: CONTESTO_NON_VALIDO, messaggio: Codici incoerenti con il percorso scelto}}
            campi: {value: {codice: CAMPO_NON_AMMESSO, messaggio: Campi duplicati nella versione}}
...